refactor: simplify direct-driver path
