+++ lib/server.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var DirectDriver = require('./drivers/direct/direct-driver');
+var DirectDriver = require('./drivers/direct');
 var Environment = require('./env');
 var EventEmitter = require('events').EventEmitter;
 var MeshServer = require('strong-mesh-models').meshServer;
